Fix negative info()['map_addr'] on 64-bit Windows under cffi
The cffi implementation converted me_mapaddr through C 'long'.  On
Windows, long is 32 bits even in 64-bit builds (LLP64), so the map
address was truncated to 32 bits and then read as signed: any address
with bit 31 set came back negative, and test_info's 'assert info[k] >= 0'
failed.  ASLR decides whether that bit is set, which is why this
surfaced as an intermittent Windows CI failure rather than a consistent
one.

Cast through uintptr_t instead, which is pointer-sized everywhere.  This
also makes cffi agree with the CPython implementation, which already
converts via intptr_t and emits an unsigned long long.

The defect is pre-existing -- the line is unchanged from master -- and
cannot reproduce on Linux, where long is 64-bit and me_mapaddr is NULL
besides (LMDB only sets it under MDB_FIXEDMAP).  Add a regression test
that exercises the conversion directly with a full-width pointer, since
neither a real mapping nor the host platform would expose it.

+++ lmdb/cffi.py
@@ -1336,7 +1336,13 @@ class Environment:
             if rc:
                 raise _error("mdb_env_info", rc)
         return {
-            "map_addr": int(_ffi.cast('long', info.me_mapaddr)),
+            # uintptr_t, not long: on Windows long is 32 bits even in
+            # 64-bit builds, so casting a map address through it truncates
+            # and then reads the result as signed, yielding a negative
+            # address whenever bit 31 happens to be set.  This matches what
+            # the CPython implementation returns, which converts via
+            # intptr_t and emits an unsigned long long.
+            "map_addr": int(_ffi.cast('uintptr_t', info.me_mapaddr)),
             "map_size": info.me_mapsize,
             "last_pgno": info.me_last_pgno,
             "last_txnid": info.me_last_txnid,
